refactor(api): merge manager related types
